ts migration(controllers typed)

+++ server/controllers/expenses-controller.ts
@@ -1,9 +1,22 @@
-// @ts-nocheck
 import Expenses from '../models/expenses-model';
 import MonthController from './month-controller';
+import {MonthCode} from 'date-types';
+import {MongoResponse} from './controller-types';
+import {IExpenseRaw} from 'expenses-types';
+
+interface IExpenseInput {
+    cash: number;
+    cashless: number;
+    monthCode: MonthCode;
+    invoice: boolean;
+    category: string;
+    description: string;
+    date: string;
+}
 
 class ExpensesController {
-    static async addNewExpense({ExpenseInput}) {
+
+    static async addNewExpense({ExpenseInput}: { ExpenseInput: IExpenseInput }): Promise<MongoResponse<IExpenseRaw>> {
         try {
             const month = await MonthController.getMonthByCode(ExpenseInput.monthCode);
 
@@ -27,7 +40,10 @@ class ExpensesController {
         }
     }
 
-    static async editExpense({ExpenseID, ExpenseInput}) {
+    static async editExpense({
+                                 ExpenseID,
+                                 ExpenseInput
+                             }: { ExpenseID: string, ExpenseInput: IExpenseInput }): Promise<MongoResponse<IExpenseRaw>> {
         const expense = {
             cash: ExpenseInput.cash,
             cashless: ExpenseInput.cashless,
@@ -37,17 +53,22 @@ class ExpensesController {
         };
 
         try {
-            return await Expenses.findByIdAndUpdate(ExpenseID, expense, {
+            const updatedExpense = await Expenses.findByIdAndUpdate(ExpenseID, expense, {
                 new: true
             });
+            if (!updatedExpense) {
+                throw new Error('Expense not found');
+            }
+            return updatedExpense;
         } catch (error) {
             throw error;
         }
     }
 
-    static async deleteExpense({ExpenseID}) {
+    static async deleteExpense({ExpenseID}: { ExpenseID: string }): Promise<'success'> {
         try {
             const expense = await Expenses.findById(ExpenseID);
+            if (!expense) throw new Error('Expense not found');
             await expense.delete();
 
             return 'success';
@@ -56,7 +77,7 @@ class ExpensesController {
         }
     }
 
-    static async getExpensesByMonth({monthCode}) {
+    static async getExpensesByMonth({monthCode}: { monthCode: MonthCode }) {
         try {
             const month = await MonthController.getMonthByCode(monthCode);
             return await Expenses.find({month: month}).sort({date: 1});
